% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-withdoc.R
\name{get_alcohol}
\alias{get_alcohol}
\title{Computation of pure alcohol}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{named numeric vector.
Names of the vector represent the consumed type of drink, the numbers represent the consumed quantitiy of the corresponding drink.
Currently the following drinks are supported: "massn", "hoibe", "wein" and "schnaps"}
}
\value{
a numeric value representing the mass of absolute alcohol in gram
}
\description{
Takes a list of alcoholic drinks and computes the cummulated mass of pure alcohol
}
\examples{
\dontrun{
get_alcohol(c("hoibe" = 3, "wein" = 2, "massn" = 1))
}
}
\references{
https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/
}
\author{
Fabian Scheipl, Marc Johler
}
