% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-withdoc.R
\name{get_bodywater}
\alias{get_bodywater}
\title{Computation of body water}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{sex of the individual. Either "male" or "female"}

\item{age}{age of the individual in years}

\item{height}{height of the individual in centimeters}

\item{weight}{weight of the individual in kilograms}
}
\value{
a numeric value representing the quantity of body water in liters
}
\description{
Computes an individuals quantity of body water
}
\examples{
\dontrun{
get_bodywater(
  age = 17,
  sex = "male",
  height = 170,
  weight = 65
)
}
}
\references{
https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/
}
\author{
Fabian Scheipl, Marc Johler
}
