% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-withdoc.R
\name{get_permille}
\alias{get_permille}
\title{Low-level function to compute blood-alcohol-level}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{mass of consumed alcohol in gram}

\item{bodywater}{quantity of individuals body water in liters}

\item{drinking_time}{two element POSIXct-vector with the first element being the start time of drinking and the second element being the end time of drinking}
}
\value{
a numeric value representing the blood-alcohol-level in per mille
}
\description{
Computes the "alcohol per mille" value based on consumed alcohol, total body water and passed time since start of drinking
}
\examples{
\dontrun{
get_permille(
  alcohol_drunk = 48,
  bodywater = 42,
  drinking_time = as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00"))
)
}
}
\references{
https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/
}
\author{
Fabian Scheipl, Marc Johler
}
