% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-withdoc.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\alias{how_drunk}
\title{Computation of blood-alcohol-level}
\usage{
tell_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{age of the individual in years}

\item{sex}{sex of the individual. Either "male" or "female"}

\item{height}{height of the individual in centimeters}

\item{weight}{weight of the individual in kilograms}

\item{drinking_time}{two element POSIXct-vector with the first element being the start time of drinking and the second element being the end time of drinking}

\item{drinks}{named numeric vector.
Names of the vector represent the consumed type of drink, the numbers represent the consumed quantitiy of the corresponding drink.
Currently the following drinks are supported: "massn", "hoibe", "wein" and "schnaps"}
}
\value{
a numeric value representing the blood-alcohol-level in per mille
}
\description{
Applies the formula for total body water according to Whatson, in order to compute the blood-alcohol-level according to Widmark.
The individuals attributes as height, weight, age and gender are included, as well as the consumed amount of alcohol and the duration of consumption.
}
\examples{
# This is supposed to work without warnings:
\dontrun{
tell_me_how_drunk(
  age = 39,
  sex = "male",
  height = 190,
  weight = 87,
  drinking_time = as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00")),
  drinks = c("massn" = 3, "schnaps" = 4)
)

# Violations of German laws according alcohol consumption will be reported as a warning:
tell_me_how_drunk(
  age = 17,
  sex = "male",
  height = 170,
  weight = 65,
  drinking_time = as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00")),
  drinks = c("wein" = 3, "schnaps" = 4)
)

tell_me_how_drunk(
  age = 15,
  sex = "female",
  height = 165,
  weight = 58,
  drinking_time = as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00")),
  drinks = c("hoibe" = 3)
)
}
}
\references{
https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/
}
\author{
Fabian Scheipl, Marc Johler
}
